+++ R/datasets.R
@@ -2,7 +2,8 @@
 
 #' MNIST database of handwritten digits
 #'
-#' Dataset of 60,000 28x28 grayscale images of the 10 digits, along with a test set of 10,000 images.
+#' Dataset of 60,000 28x28 grayscale images of the 10 digits, along with a test
+#' set of 10,000 images.
 #'
 #' @param ntrain number of training samples
 #' @param ntest number of test samples
@@ -10,7 +11,7 @@
 #' @family datasets
 #'
 #' @export
-dataset_mnist <- function(ntrain = 60000L, ntest = 10000L, onehot = TRUE) {
+dataset_mnist_digits <- function(ntrain = 60000L, ntest = 10000L, onehot = TRUE) {
     # import data_utils module
     python_path <- system.file("python", package = "rTorch")
     tools <- import_from_path("torchtools", path = python_path)