+++ R/generics.R
@@ -38,6 +38,17 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
 }
 
 
+#' One tensor operation
+#'
+#' @param x tensor
+#' @examples
+#' A <- torch$ones(c(60000L, 1L, 28L, 28L))
+#' dim(A)
+one_tensor_op <- function(x) UseMethod("one_tensor_op")
+
+
+#' @describeIn one_tensor_op Dimensions of a tensor
+#' @details Get the dimensions of a tensor displaying it as a vector.
 #' @export
 "dim.torch.Tensor" <- function(x) {        # change .tensor to .Tensor
     if (py_is_null_xptr(x))
@@ -52,7 +63,7 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
     }
 }
 
-
+#' @describeIn one_tensor_op Length of a tensor. Eqivalent to torch$numel()
 #' @export
 "length.torch.Tensor" <- function(x) {
     if (py_is_null_xptr(x))
@@ -62,34 +73,131 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
 }
 
 
+#' Remainder
+#'
+#' Computes the element-wise remainder of division.
+#' @param a a tensor
+#' @param b a scalar or a tensor
+#' @export
+#' @examples
+#' x <- torch$Tensor(list(-3., -2, -1, 1, 2, 3))
+#' y <- torch$Tensor(list(1., 2, 3, 4, 5))
+#' torch$remainder(x, 2)
+#' torch$remainder(y, 1.5)
+#'
+#' x %% 2
+#' y %% 1.5
+#'
+"%%.torch.Tensor" <- function(a, b) {
+    torch$remainder(a, b)
+}
+
+
 # all <- function(x, ...) UseMethod("all")
 # any <- function(x, ...) UseMethod("any")
 
 
+#' all
+#'
+#' Returns True if all elements in the tensor are non-zero, False otherwise.
+#' @param x tensor
+#' @param dim dimension to reduce
+#' @param ... other parameters (yet to be developed)
 #' @export
-"all.torch.Tensor" <- function(x, ...) {
+#' @examples
+#' a <- torch$BoolTensor(list(TRUE, TRUE, TRUE, TRUE))
+#' b <- torch$BoolTensor(list(FALSE, TRUE, TRUE, TRUE))
+#' c <- torch$BoolTensor(list(TRUE, TRUE, TRUE, FALSE))
+#' all(a)
+#' all(b)
+#' all(c)
+#' d <- torch$tensor(list(list(0, 0),
+#'                        list(0, 0),
+#'                        list(0, 1),
+#'                        list(1, 1)), dtype=torch$uint8)
+#' all(d)
+#' all(d, dim=0L)
+#' all(d, dim=1L)
+"all.torch.Tensor" <- function(x, dim, ...) {
     # quick version of torch$all
     # TODO: modify to use all arguments
     # all(dim, keepdim=False, out=None) → Tensor
     # DO NOT USE torch$tensor() to prevent warning:
     #            ... it is recommended to use sourceTensor.clone().detach()
     x <- torch$as_tensor(x, dtype = torch$uint8)
-    as.logical(torch$all(x)$numpy())
+    # as.logical(torch$all(x)$numpy())
+    if (missing(dim)) torch$all(x) else torch$all(x, dim=as.integer(dim))
 }
 
+
+#' any
+#'
+#' Returns True if any elements in the tensor are non-zero, False otherwise.
+#' @param x tensor
+#' @param dim dimension to reduce
+#' @param ... other params (yet to be developed)
 #' @export
-"any.torch.Tensor" <- function(x, ...) {
+#' @examples
+#' a <- torch$BoolTensor(list(TRUE, TRUE, TRUE, TRUE))
+#' b <- torch$BoolTensor(list(FALSE, TRUE, TRUE, TRUE))
+#' c <- torch$BoolTensor(list(TRUE, TRUE, TRUE, FALSE))
+#' any(a)
+#' any(b)
+#' any(c)
+#' d <- torch$tensor(list(list(1, 0),
+#'                        list(0, 0),
+#'                        list(0, 1),
+#'                        list(0, 0)), dtype=torch$uint8)
+#' any(d)
+#' any(d, dim=0L)
+#' any(d, dim=1L)
+"any.torch.Tensor" <- function(x, dim, ...) {
     # quick version of torch$any
     # TODO: modify to use all arguments
     # all(dim, keepdim=False, out=None) → Tensor
     # DO NOT USE torch$tensor() to prevent warning:
     #            ... it is recommended to use sourceTensor.clone().detach()
     x <- torch$as_tensor(x, dtype = torch$uint8)
-    as.logical(torch$any(x)$numpy())
+    # as.logical(torch$any(x)$numpy())
+    if (missing(dim)) torch$any(x) else torch$any(x, dim=as.integer(dim))
 }
 
 
+#' Two tensor operations
+#'
+#' @param a tensor
+#' @param b tensor
+#' @examples
+#' a <- torch$Tensor(list(1, 1, 1))
+#' b <- torch$Tensor(list(2, 2, 2))
+#' s <- 2.0
+#' a + b
+#' b - a
+#' a * b
+#' a / s
+#' a == b
+#' a == a
+#' a != a
+#' x <- torch$Tensor(list(list(2, 2, 2), list(4, 4, 4)))
+#' y <- torch$Tensor(list(list(1, 2, 1), list(3, 4, 5)))
+#' x > y
+#' x < y
+#' x >= y
+#' y <= x
+#' diag <- torch$eye(3L)
+#' zeros <- torch$zeros(c(3L, 3L))
+#' diag & zeros
+#' diag & diag
+#' diag | diag
+#' zeros | zeros
+#' zeros & zeros
+#' diag & zeros
+#' diag | zeros
+tensor_ops <- function(a, b) UseMethod("tensor_ops")
 
+
+
+#' @describeIn tensor_ops Add two tensors
 #' @export
 "+.torch.Tensor" <- function(a, b) {
     if (any(class(a) == "torch.Tensor"))
@@ -99,6 +207,7 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
 }
 
 
+#' @describeIn tensor_ops Subtract two tensors
 #' @export
 "-.torch.Tensor" <- function(a, b) {
     if (missing(b)) {
@@ -115,7 +224,9 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
 }
 
 
+
 #' @export
+#' @describeIn tensor_ops Tensor multiplication
 "*.torch.Tensor" <- function(a, b) {
     if (py_has_attr(torch, "multiply"))
         torch$multiply(a, b)
@@ -123,6 +234,7 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
         torch$mul(a, b)
 }
 
+#' @describeIn tensor_ops Divide two tensors
 #' @export
 "/.torch.Tensor" <- function(a, b) {
     torch$div(a, b)
@@ -131,28 +243,7 @@ py_str.torch.python.ops.variables.Variable <- function(object, ...) {
 
 
 
-
-# TODO: finish these two and tensor float ###################
-#' @export
-"*.torch.Tensor" <- function(a, b) {
-    if (any(class(a) == "torch._C.FloatTensorBase"))
-        torch$mul(a, b)
-    else
-        torch$mul(b, a)
-}
-
-
-#' @export
-"/.torch.Tensor" <- function(a, b) {
-    if (any(class(a) == "torch._C.FloatTensorBase"))
-        torch$div(a, b)
-    else
-        torch$div(b, a)
-}
-
-##################################################################
-
-
+#' @describeIn tensor_ops Compare two tensors if equal
 #' @export
 "==.torch.Tensor" <- function(a, b) {
     torch$as_tensor(torch$eq(a, b), dtype = torch$bool)
@@ -167,48 +258,57 @@ tensor_not_equal <- function(x, y) {
     torch$BoolTensor(np$not_equal(x, y))
 }
 
+#' @describeIn tensor_ops Compare two tensors if not equal
 #' @export
 "!=.torch.Tensor" <- function(a, b) {
     # there is not not_equal function in PyTorch
     # tensor_not_equal(a, b)
-    torch$ne(a, b)
+    # torch$ne(a, b)
+    torch$as_tensor(torch$ne(a, b), dtype = torch$bool)
 }
 
+#' @describeIn one_tensor_op Logical NOT of a tensor
 #' @export
-"!.torch.Tensor" <- function(a) {
+"!.torch.Tensor" <- function(x) {
     # there is not logical not in torch
     # torch$BoolTensor(np$logical_not(a))
-    torch$as_tensor(np$logical_not(a), dtype = torch$bool)
+    torch$as_tensor(np$logical_not(x), dtype = torch$bool)
 }
 
 
-
+#' @describeIn tensor_ops A tensor less than another tensor
 #' @export
 "<.torch.Tensor" <- function(a, b) {
-    torch$lt(a, b)
+    # torch$lt(a, b)
+    torch$as_tensor(torch$lt(a, b), dtype = torch$bool)
 }
 
 
+#' @describeIn tensor_ops A tensor less or equal than another tensor
 #' @export
 "<=.torch.Tensor" <- function(a, b) {
-    torch$le(a, b)
+    # torch$le(a, b)
+    torch$as_tensor(torch$le(a, b), dtype = torch$bool)
 }
 
 
+#' @describeIn tensor_ops A tensor greater than another tensor
 #' @export
 ">.torch.Tensor" <- function(a, b) {
-    torch$gt(a, b)
+    # torch$gt(a, b)
+    torch$as_tensor(torch$gt(a, b), dtype = torch$bool)
 }
 
-
+#' @describeIn tensor_ops A tensor greater or equal than another tensor
 #' @export
 ">=.torch.Tensor" <- function(a, b) {
-    torch$ge(a, b)
+    # torch$ge(a, b)
+    torch$as_tensor(torch$ge(a, b), dtype = torch$bool)
 }
 
 
 
-#' Dot product
+#' Dot product of two tensors
 #' PyTorch dot function
 #' @param a Tensor 1
 #' @param b Tensor 2
@@ -218,7 +318,7 @@ tensor_not_equal <- function(x, y) {
 }
 
 
-#' Matrix/Tensor multiplication
+#' Matrix/Tensor multiplication of two tensors
 #' PyTorch matmul
 #' @param a Tensor 1
 #' @param b Tensor 2
@@ -230,18 +330,23 @@ tensor_not_equal <- function(x, y) {
 
 
 
-
-
+#' @describeIn tensor_ops A tensor 'a' to the power of 'b'
 #' @export
 "^.torch.Tensor" <- function(a, b) {
     torch$pow(a, b)
 }
 
+
+#' @describeIn one_tensor_op Exponential of a tensor
 #' @export
 "exp.torch.Tensor" <- function(x) {
     torch$exp(x)
 }
 
+
+#' Logarithm of a tensor given the tensor and the base
+#' @param x a tensor
+#' @param base the base of the logarithm
 #' @export
 "log.torch.Tensor" <- function(x, base = exp(1L)) {
     if (is_tensor(base) || base != exp(1L)) {
@@ -252,16 +357,18 @@ tensor_not_equal <- function(x, y) {
         # print("not here")
         torch$log(x)
     }
-
-
 }
 
+#' Logarithm of a tensor in base 2
+#' @param x a tensor
 #' @export
 #' @method log2 torch.Tensor
 "log2.torch.Tensor" <- function(x) {
     torch$log2(x)
 }
 
+#' Logarithm of a tensor in base 10
+#' @param x a tensor
 #' @export
 #' @method log10 torch.Tensor
 "log10.torch.Tensor" <- function(x) {
@@ -283,12 +390,14 @@ tensor_logical_or <- function(x, y) {
 }
 
 
+#' @describeIn tensor_ops Logical AND of two tensors
 #' @export
 "&.torch.Tensor" <- function(a, b) {
     tensor_logical_and(a, b)
 }
 
 
+#' @describeIn tensor_ops Logical OR of two tensors
 #' @export
 "|.torch.Tensor" <- function(a, b) {
     tensor_logical_or(a, b)