+++ tests/testthat/test_generics.R
@@ -3,6 +3,9 @@ library(testthat)
 
 source("tensor_functions.R")
 
+TRUE_TENSOR  <- torch$as_tensor(1L, dtype=torch$uint8)
+FALSE_TENSOR <- torch$as_tensor(0L, dtype=torch$uint8)
+
 all_boolean <- function(x) {
   # convert tensor of 1s and 0s to a unique boolean
   as.logical(torch$all(x)$numpy())
@@ -12,6 +15,13 @@ expect_all_equal <- function(x, y) {
   all_boolean(torch$eq(x, y))
 }
 
+expect_false_tensor <- function(x) {
+  expect_equal(x, FALSE_TENSOR)
+}
+
+expect_true_tensor <- function(x) {
+  expect_equal(x, TRUE_TENSOR)
+}
 
 context("dim on tensor")
 
@@ -109,8 +119,10 @@ test_that("tensor equal as x == y", {
       list(1, 1, 1, 1),
       list(1, 1, 1, 1)
     ))
-    expect_true(all(torch$eq(x, y)))
-    expect_true(all(x == y))
+    # true_tensor <- torch$as_tensor(1L, dtype=torch$uint8)
+    # print(true_tensor)
+    expect_equal(all(torch$eq(x, y)), TRUE_TENSOR)
+    expect_equal(all(x == y), TRUE_TENSOR)
     expect_equal((x == y), expected)
 })
 
@@ -120,33 +132,32 @@ context("tensors not equal !=")
 test_that("tensor equal as x == y", {
   A <- torch$ones(60000L, 3L, 28L, 28L)
   B <- torch$ones(60000L, 3L, 28L, 28L)
-  expect_true(all(A == B))
-  expect_false(all(A != B))
+  expect_equal(all(A == B), TRUE_TENSOR)
+  expect_equal(all(A != B), TRUE_TENSOR)
 })
 
 
 test_that("tensor not equal as x != y", {
   a <- torch$Tensor(list(2, 2, 2, 2))
   b <- torch$Tensor(list(2, 2, 2, 1))
-  expect_false(all(a == b))
-  expect_true(any(a != b))
-  # print((a != b))
-  expect_equal((a != b), torch$BoolTensor(list(FALSE, FALSE, FALSE, TRUE)))
+  expect_equal(all(a == b), FALSE_TENSOR)
+  expect_equal(any(a != b), TRUE_TENSOR)
+  # expect_equal((a != b), torch$BoolTensor(list(FALSE, FALSE, FALSE, TRUE)))
 
   x = torch$rand(5L, 4L)
   y = torch$rand(5L, 4L)
-  expect_false(all(torch$eq(x, y)))
-  expect_true(!all(torch$eq(x, y)))
-  expect_true(all(x != y))
+  expect_false_tensor(all(torch$eq(x, y)))
+  expect_true_tensor(!all(torch$eq(x, y)))
+  expect_true_tensor(all(x != y))
 
 
   A <- torch$ones(60000L, 3L, 28L, 28L)    # all ones
   B <- torch$zeros(60000L, 3L, 28L, 28L)   # all zeroes
-  expect_true(all(A != B))
+  expect_true_tensor(all(A != B))
 
   A <- torch$ones(60000L, 1L, 28L, 28L)
   B <- torch$ones(60000L, 1L, 28L, 28L)
-  expect_false(all(A != B))
+  expect_false_tensor(all(A != B))
 })
 
 
@@ -154,19 +165,19 @@ test_that("tensor not equal as x != y", {
 
 test_that("numpy logical-not works as !", {
   A <- torch$ones(5L)
-  expect_false(all(!A))
+  expect_false_tensor(all(!A))
 
   Z <- torch$zeros(5L)
-  expect_false(all(Z))
-  expect_true(all(!Z))
+  expect_false_tensor(all(Z))
+  expect_true_tensor(all(!Z))
 
   E <- torch$eye(3L)
-  expect_true(all(torch$diag(E)))
+  expect_true_tensor(all(torch$diag(E)))
 
   # torch$diag does not work on boolean tensors when diag are all zeros
   # the tensor has to be converted first to uint8 dtype
   NE <- torch$as_tensor(!E, dtype = torch$uint8)
-  expect_false(all(NE$diag()))
+  expect_false_tensor(all(NE$diag()))
 
   expect_error(torch$diag(!E))    # this will throw an error
 })
@@ -178,9 +189,9 @@ context("tensor comparison")
 test_that("tensor is less than ", {
   A <- torch$ones(60000L, 1L, 28L, 28L)
   C <- A * 0.5
-  expect_true(all_boolean(torch$lt(C, A)))
-  expect_true(all_boolean(C < A))
-  expect_false(all_boolean(A < C))
+  expect_true_tensor(all(torch$lt(C, A)))
+  expect_true_tensor(all(C < A))
+  expect_false_tensor(all(A < C))
 })
 
 
@@ -188,20 +199,20 @@ test_that("tensor is greater than ", {
   A <- torch$ones(60000L, 1L, 28L, 28L)
   D <- A * 2.0
   # print(as.logical(torch$all(torch$gt(D, A))$numpy()))
-  expect_true(all_boolean(torch$gt(D, A)))
-  expect_false(all_boolean(torch$gt(A, D)))
+  expect_true_tensor(all(torch$gt(D, A)))
+  expect_false_tensor(all(torch$gt(A, D)))
 })
 
 test_that("tensor is less than or equal", {
   A <- torch$ones(60000L, 1L, 28L, 28L)
-  expect_true(all_boolean(torch$le(A, A)))
-  expect_true(all_boolean(A <= A))
+  expect_true_tensor(all(torch$le(A, A)))
+  expect_true_tensor(all(A <= A))
 })
 
 test_that("tensor is greater than or equal", {
   A <- torch$ones(60000L, 1L, 28L, 28L)
-  expect_true(all_boolean(torch$ge(A, A)))
-  expect_true(all_boolean(A >= A))
+  expect_true_tensor(all(torch$ge(A, A)))
+  expect_true_tensor(all(A >= A))
 })