+++ NEWS.md
@@ -1,3 +1,12 @@
+# rTorch 0.0.1.9013
+* July 25 2019
+* Add function remainder for tensors. Equivalent to `a %% b`
+* Change unit tests in `test_generics.R` to use new function `expect_true_tensor`
+* Enhance functions `any` and `all`. Add examples
+* Add roxygen documentation to two tensor operations
+* Change download folders for MNIST datasets under project folder
+
+
 # rTorch 0.0.1.9012
 * July 24 2019
 * Change MNIST download folder to ~/raw_data instead of inst/